Next step towards long int support for translation indices. Replaced mainly "int" by "long" in operator calls.

+++ lawa/operators/pdeoperators1d/identityoperator1d.tcc
@@ -8,8 +8,8 @@ IdentityOperator1D<T, Basis>::IdentityOperator1D(const Basis& _basis)
 
 template <typename T, typename Basis>
 T
-IdentityOperator1D<T, Basis>::operator()(XType xtype1, int j1, int k1,
-                                         XType xtype2, int j2, int k2) const
+IdentityOperator1D<T, Basis>::operator()(XType xtype1, int j1, long k1,
+                                         XType xtype2, int j2, long k2) const
 {   
     return integral(j1, k1, xtype1, 0, j2, k2, xtype2, 0);
 }
