+++ lawa/operators/pdeoperators1d/identityoperator1d_pg.tcc
@@ -8,8 +8,8 @@ IdentityOperator1D_PG<T, TrialBasis, TestBasis>::IdentityOperator1D_PG(const Tri
 
 template <typename T, typename TrialBasis, typename TestBasis>
 T
-IdentityOperator1D_PG<T, TrialBasis, TestBasis>::operator()(XType xtype1, int j1, int k1,
-                                                            XType xtype2, int j2, int k2) const
+IdentityOperator1D_PG<T, TrialBasis, TestBasis>::operator()(XType xtype1, int j1, long k1,
+                                                            XType xtype2, int j2, long k2) const
 {   
     return integral(j1, k1, xtype1, 0, j2, k2, xtype2, 0);
 }
