+++ lawa/operators/pdeoperators2d/helmholtzoperator2d.tcc
@@ -28,10 +28,10 @@ HelmholtzOperator2D<T, Basis2D>::HelmholtzOperator2D(const Basis2D &_basis, cons
 
 template <typename T, typename Basis2D>
 T
-HelmholtzOperator2D<T, Basis2D>::operator()(XType row_xtype_x, int j1_x, int k1_x,
-                                            XType row_xtype_y, int j1_y, int k1_y,
-                                            XType col_xtype_x, int j2_x, int k2_x,
-                                            XType col_xtype_y, int j2_y, int k2_y) const
+HelmholtzOperator2D<T, Basis2D>::operator()(XType row_xtype_x, int j1_x, long k1_x,
+                                            XType row_xtype_y, int j1_y, long k1_y,
+                                            XType col_xtype_x, int j2_x, long k2_x,
+                                            XType col_xtype_y, int j2_y, long k2_y) const
 {   
     // (v1_x * u1_x) * (v2 * u2) + (v1 * u1) * (v2_y * u2_y) + c * (v1 * u1) * (v2 * u2)
     return integral_x(j1_x, k1_x, row_xtype_x, 1, j2_x, k2_x, col_xtype_x, 1)
