+++ lawa/preconditioners/preconditioners1d/H1normpreconditioner1d.h
@@ -36,7 +36,7 @@ class H1NormPreconditioner1D
         H1NormPreconditioner1D(const Basis &_basis);
 
         T
-        operator()(XType xtype1, int j1, int k1) const;
+        operator()(XType xtype1, int j1, long k1) const;
 
         T
         operator()(const Index1D &index) const;
