+++ lawa/preconditioners/preconditioners1d/H1normpreconditioner1d.tcc
@@ -10,7 +10,7 @@ H1NormPreconditioner1D<T,Basis>::H1NormPreconditioner1D(const Basis &basis)
 
 template <typename T, typename Basis>
 T
-H1NormPreconditioner1D<T,Basis>::operator()(XType xtype, int j, int k) const
+H1NormPreconditioner1D<T,Basis>::operator()(XType xtype, int j, long k) const
 {
     return 1./std::sqrt(_integral(j,k,xtype,0,j,k,xtype,0)
                       + _integral(j,k,xtype,1,j,k,xtype,1));
