+++ lawa/righthandsides/rhs2d.h
@@ -31,8 +31,8 @@ struct Rhs2D {
     operator()(const Index2D &index) const = 0;
     
     virtual T
-    operator()(XType xtype_x, int j_x, int k_x,
-               XType xtype_y, int j_y, int k_y) const = 0;
+    operator()(XType xtype_x, int j_x, long k_x,
+               XType xtype_y, int j_y, long k_y) const = 0;
 };
     
 } // namespace lawa
